+++ flake.nix
@@ -43,6 +43,32 @@
           SHADERC_LIB_DIR = lib.makeLibraryPath [ shaderc ];
           VK_LAYER_PATH = "${vulkan-validation-layers}/share/vulkan/explicit_layer.d";
         };
+        devShells.CI = with pkgs; mkShell rec {
+          buildInputs = [
+            (rust-bin.stable.latest.minimal.override {
+              extensions = [ "clippy" ];
+              # Windows CI unfortunately needs to cross-compile from within WSL because Nix doesn't
+              # work on Windows.
+              targets = [ "x86_64-pc-windows-msvc" ];
+            })
+            # We use nightly rustfmt features.
+            (rust-bin.selectLatestNightlyWith (toolchain: toolchain.rustfmt))
+
+            # Vulkan dependencies
+            shaderc
+
+            # winit dependencies
+            libxkbcommon
+            wayland
+            xorg.libX11
+            xorg.libXcursor
+            xorg.libXi
+            xorg.libXrandr
+          ];
+
+          LD_LIBRARY_PATH = lib.makeLibraryPath buildInputs;
+          SHADERC_LIB_DIR = lib.makeLibraryPath [ shaderc ];
+        };
       }
     );
 }