+++ vulkano/src/command_buffer/commands/copy.rs
@@ -701,12 +701,12 @@ impl RecordingCommandBuffer {
             .validate(self.device())
             .map_err(|err| err.add_context("copy_image_info"))?;
 
-        let &CopyImageInfo {
-            ref src_image,
+        let CopyImageInfo {
+            src_image,
             src_image_layout: _,
-            ref dst_image,
+            dst_image,
             dst_image_layout: _,
-            ref regions,
+            regions,
             _ne: _,
         } = copy_image_info;
 
@@ -1780,10 +1780,10 @@ impl CopyBufferInfo {
     }
 
     pub(crate) fn validate(&self, device: &Device) -> Result<(), Box<ValidationError>> {
-        let &Self {
-            ref src_buffer,
-            ref dst_buffer,
-            ref regions,
+        let Self {
+            src_buffer,
+            dst_buffer,
+            regions,
             _ne: _,
         } = self;
 
@@ -1906,9 +1906,9 @@ impl CopyBufferInfo {
         &self,
         regions_vk: &'a [vk::BufferCopy2<'static>],
     ) -> vk::CopyBufferInfo2<'a> {
-        let &Self {
-            ref src_buffer,
-            ref dst_buffer,
+        let Self {
+            src_buffer,
+            dst_buffer,
             regions: _,
             _ne: _,
         } = self;
@@ -1920,10 +1920,10 @@ impl CopyBufferInfo {
     }
 
     pub(crate) fn to_vk2_regions(&self) -> SmallVec<[vk::BufferCopy2<'static>; 8]> {
-        let &Self {
-            ref src_buffer,
-            ref dst_buffer,
-            ref regions,
+        let Self {
+            src_buffer,
+            dst_buffer,
+            regions,
             _ne: _,
         } = self;
 
@@ -1939,9 +1939,9 @@ impl CopyBufferInfo {
     }
 
     pub(crate) fn to_vk(&self) -> CopyBufferInfoVk {
-        let &Self {
-            ref src_buffer,
-            ref dst_buffer,
+        let Self {
+            src_buffer,
+            dst_buffer,
             regions: _,
             _ne: _,
         } = self;
@@ -1953,10 +1953,10 @@ impl CopyBufferInfo {
     }
 
     pub(crate) fn to_vk_regions(&self) -> SmallVec<[vk::BufferCopy; 8]> {
-        let &Self {
-            ref src_buffer,
-            ref dst_buffer,
-            ref regions,
+        let Self {
+            src_buffer,
+            dst_buffer,
+            regions,
             _ne: _,
         } = self;
 
@@ -3481,7 +3481,7 @@ pub(crate) struct CopyImageInfoVk {
 pub struct ImageCopy {
     /// The subresource of `src_image` to copy from.
     ///
-    /// The default value is empty, which must be overridden.
+    /// The default value is [`ImageSubresourceLayers::default()`].
     pub src_subresource: ImageSubresourceLayers,
 
     /// The offset from the zero coordinate of `src_image` that copying will start from.
@@ -3491,7 +3491,7 @@ pub struct ImageCopy {
 
     /// The subresource of `dst_image` to copy to.
     ///
-    /// The default value is empty, which must be overridden.
+    /// The default value is [`ImageSubresourceLayers::default()`].
     pub dst_subresource: ImageSubresourceLayers,
 
     /// The offset from the zero coordinate of `dst_image` that copying will start from.
@@ -3519,19 +3519,9 @@ impl ImageCopy {
     #[inline]
     pub const fn new() -> Self {
         Self {
-            src_subresource: ImageSubresourceLayers {
-                aspects: ImageAspects::empty(),
-                mip_level: 0,
-                base_array_layer: 0,
-                layer_count: 0,
-            },
+            src_subresource: ImageSubresourceLayers::new(),
             src_offset: [0; 3],
-            dst_subresource: ImageSubresourceLayers {
-                aspects: ImageAspects::empty(),
-                mip_level: 0,
-                base_array_layer: 0,
-                layer_count: 0,
-            },
+            dst_subresource: ImageSubresourceLayers::new(),
             dst_offset: [0; 3],
             extent: [0; 3],
             _ne: crate::NE,
@@ -5092,7 +5082,7 @@ pub struct BufferImageCopy {
 
     /// The subresource of the image to copy from/to.
     ///
-    /// The default value is empty, which must be overridden.
+    /// The default value is [`ImageSubresourceLayers::default()`].
     pub image_subresource: ImageSubresourceLayers,
 
     /// The offset from the zero coordinate of the image that copying will start from.
@@ -5123,12 +5113,7 @@ impl BufferImageCopy {
             buffer_offset: 0,
             buffer_row_length: 0,
             buffer_image_height: 0,
-            image_subresource: ImageSubresourceLayers {
-                aspects: ImageAspects::empty(),
-                mip_level: 0,
-                base_array_layer: 0,
-                layer_count: 0,
-            },
+            image_subresource: ImageSubresourceLayers::new(),
             image_offset: [0; 3],
             image_extent: [0; 3],
             _ne: crate::NE,
@@ -5138,7 +5123,7 @@ impl BufferImageCopy {
     // Following
     // https://registry.khronos.org/vulkan/specs/1.3-extensions/html/chap20.html#copies-buffers-images-addressing
     pub(crate) fn buffer_copy_size(&self, format: Format) -> DeviceSize {
-        let &BufferImageCopy {
+        let &Self {
             buffer_offset: _,
             mut buffer_row_length,
             mut buffer_image_height,
@@ -6208,7 +6193,7 @@ pub(crate) struct BlitImageInfoVk {
 pub struct ImageBlit {
     /// The subresource of `src_image` to blit from.
     ///
-    /// The default value is empty, which must be overridden.
+    /// The default value is [`ImageSubresourceLayers::default()`].
     pub src_subresource: ImageSubresourceLayers,
 
     /// The offsets from the zero coordinate of `src_image`, defining two corners of the region
@@ -6221,7 +6206,7 @@ pub struct ImageBlit {
 
     /// The subresource of `dst_image` to blit to.
     ///
-    /// The default value is empty, which must be overridden.
+    /// The default value is [`ImageSubresourceLayers::default()`].
     pub dst_subresource: ImageSubresourceLayers,
 
     /// The offset from the zero coordinate of `dst_image` defining two corners of the
@@ -6247,19 +6232,9 @@ impl ImageBlit {
     #[inline]
     pub const fn new() -> Self {
         Self {
-            src_subresource: ImageSubresourceLayers {
-                aspects: ImageAspects::empty(),
-                mip_level: 0,
-                base_array_layer: 0,
-                layer_count: 0,
-            },
+            src_subresource: ImageSubresourceLayers::new(),
             src_offsets: [[0; 3]; 2],
-            dst_subresource: ImageSubresourceLayers {
-                aspects: ImageAspects::empty(),
-                mip_level: 0,
-                base_array_layer: 0,
-                layer_count: 0,
-            },
+            dst_subresource: ImageSubresourceLayers::new(),
             dst_offsets: [[0; 3]; 2],
             _ne: crate::NE,
         }
@@ -7050,7 +7025,7 @@ pub(crate) struct ResolveImageInfoVk {
 pub struct ImageResolve {
     /// The subresource of `src_image` to resolve from.
     ///
-    /// The default value is empty, which must be overridden.
+    /// The default value is [`ImageSubresourceLayers::default()`].
     pub src_subresource: ImageSubresourceLayers,
 
     /// The offset from the zero coordinate of `src_image` that resolving will start from.
@@ -7060,7 +7035,7 @@ pub struct ImageResolve {
 
     /// The subresource of `dst_image` to resolve into.
     ///
-    /// The default value is empty, which must be overridden.
+    /// The default value is [`ImageSubresourceLayers::default()`].
     pub dst_subresource: ImageSubresourceLayers,
 
     /// The offset from the zero coordinate of `dst_image` that resolving will start from.
@@ -7088,19 +7063,9 @@ impl ImageResolve {
     #[inline]
     pub const fn new() -> Self {
         Self {
-            src_subresource: ImageSubresourceLayers {
-                aspects: ImageAspects::empty(),
-                mip_level: 0,
-                base_array_layer: 0,
-                layer_count: 0,
-            },
+            src_subresource: ImageSubresourceLayers::new(),
             src_offset: [0; 3],
-            dst_subresource: ImageSubresourceLayers {
-                aspects: ImageAspects::empty(),
-                mip_level: 0,
-                base_array_layer: 0,
-                layer_count: 0,
-            },
+            dst_subresource: ImageSubresourceLayers::new(),
             dst_offset: [0; 3],
             extent: [0; 3],
             _ne: crate::NE,