+++ vulkano/src/memory/mod.rs
@@ -547,9 +547,8 @@ impl MemoryProperties {
     }
 
     pub(crate) fn from_vk2(val_vk: &vk::PhysicalDeviceMemoryProperties2<'_>) -> Self {
-        let &vk::PhysicalDeviceMemoryProperties2 {
-            ref memory_properties,
-            ..
+        let vk::PhysicalDeviceMemoryProperties2 {
+            memory_properties, ..
         } = val_vk;
 
         Self::from_vk(memory_properties)
@@ -946,6 +945,28 @@ impl MemoryFdProperties {
 
     pub(crate) fn from_vk(val_vk: &vk::MemoryFdPropertiesKHR<'_>) -> Self {
         let &vk::MemoryFdPropertiesKHR {
+            memory_type_bits, ..
+        } = val_vk;
+
+        Self { memory_type_bits }
+    }
+}
+
+/// The properties of a Windows handle when it is imported.
+#[derive(Clone, Debug)]
+#[non_exhaustive]
+pub struct MemoryWin32HandleProperties {
+    /// A bitmask of the indices of memory types that can be used with the handle.
+    pub memory_type_bits: u32,
+}
+
+impl MemoryWin32HandleProperties {
+    pub(crate) fn to_mut_vk() -> vk::MemoryWin32HandlePropertiesKHR<'static> {
+        vk::MemoryWin32HandlePropertiesKHR::default()
+    }
+
+    pub(crate) fn from_vk(val_vk: &vk::MemoryWin32HandlePropertiesKHR<'_>) -> Self {
+        let &vk::MemoryWin32HandlePropertiesKHR {
             memory_type_bits, ..
         } = val_vk;